requirements_render.txt
aiohttp==3.9.5
python-dotenv==1.1.0
SQLAlchemy==2.0.30
asyncpg==0.30.0
